HDDS-16247. Separate StorageContainerLocationProtocol RPC identity from its internal Java interface
