+++ hadoop-ozone/ozone-manager/src/test/java/org/apache/hadoop/ozone/om/TestScmClient.java
@@ -50,7 +50,7 @@ import org.apache.hadoop.hdds.scm.container.common.helpers.ContainerWithPipeline
 import org.apache.hadoop.hdds.scm.pipeline.Pipeline;
 import org.apache.hadoop.hdds.scm.pipeline.PipelineID;
 import org.apache.hadoop.hdds.scm.protocol.ScmBlockLocationProtocol;
-import org.apache.hadoop.hdds.scm.protocol.StorageContainerLocationProtocol;
+import org.apache.hadoop.hdds.scm.protocol.StorageContainerLocationInternalInterface;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -61,13 +61,13 @@ import org.junit.jupiter.params.provider.MethodSource;
  * ScmClient test-cases.
  */
 public class TestScmClient {
-  private StorageContainerLocationProtocol containerLocationProtocol;
+  private StorageContainerLocationInternalInterface containerLocationProtocol;
   private ScmClient scmClient;
 
   @BeforeEach
   public void setUp() {
     ScmBlockLocationProtocol scmBlockLocationProtocol = mock(ScmBlockLocationProtocol.class);
-    containerLocationProtocol = mock(StorageContainerLocationProtocol.class);
+    containerLocationProtocol = mock(StorageContainerLocationInternalInterface.class);
     OzoneConfiguration conf = new OzoneConfiguration();
     scmClient = new ScmClient(scmBlockLocationProtocol,
         containerLocationProtocol, conf);
