+++ hadoop-ozone/recon/src/main/java/org/apache/hadoop/ozone/recon/api/PendingDeletionEndpoint.java
@@ -25,7 +25,7 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.Response;
 import org.apache.hadoop.hdds.protocol.proto.HddsProtos;
-import org.apache.hadoop.hdds.scm.protocol.StorageContainerLocationProtocol;
+import org.apache.hadoop.hdds.scm.protocol.StorageContainerLocationInternalInterface;
 import org.apache.hadoop.ozone.recon.api.types.DataNodeMetricsCompleteResponse;
 import org.apache.hadoop.ozone.recon.api.types.ScmPendingDeletion;
 import org.slf4j.Logger;
@@ -42,13 +42,13 @@ public class PendingDeletionEndpoint {
   private static final Logger LOG = LoggerFactory.getLogger(PendingDeletionEndpoint.class);
   private final ReconGlobalMetricsService reconGlobalMetricsService;
   private final DataNodeMetricsService dataNodeMetricsService;
-  private final StorageContainerLocationProtocol scmClient;
+  private final StorageContainerLocationInternalInterface scmClient;
 
   @Inject
   public PendingDeletionEndpoint(
       ReconGlobalMetricsService reconGlobalMetricsService,
       DataNodeMetricsService dataNodeMetricsService,
-      StorageContainerLocationProtocol scmClient) {
+      StorageContainerLocationInternalInterface scmClient) {
     this.reconGlobalMetricsService = reconGlobalMetricsService;
     this.dataNodeMetricsService = dataNodeMetricsService;
     this.scmClient = scmClient;
