+++ hadoop-ozone/recon/src/test/java/org/apache/hadoop/ozone/recon/api/TestEndpoints.java
@@ -94,7 +94,7 @@ import org.apache.hadoop.hdds.scm.node.NodeManager;
 import org.apache.hadoop.hdds.scm.node.NodeStatus;
 import org.apache.hadoop.hdds.scm.pipeline.Pipeline;
 import org.apache.hadoop.hdds.scm.pipeline.PipelineManager;
-import org.apache.hadoop.hdds.scm.protocol.StorageContainerLocationProtocol;
+import org.apache.hadoop.hdds.scm.protocol.StorageContainerLocationInternalInterface;
 import org.apache.hadoop.hdds.scm.server.OzoneStorageContainerManager;
 import org.apache.hadoop.hdds.upgrade.HDDSLayoutVersionManager;
 import org.apache.hadoop.hdds.utils.db.Table;
@@ -194,7 +194,7 @@ public class TestEndpoints extends AbstractReconSqlDBTest {
   private static final String PROMETHEUS_TEST_RESPONSE_FILE =
       "prometheus-test-response.txt";
   private ReconUtils reconUtilsMock;
-  private StorageContainerLocationProtocol mockScmClient;
+  private StorageContainerLocationInternalInterface mockScmClient;
 
   private List<HddsProtos.Node> nodes = getNodeDetails(2);
   private Map<String, List<ContainerID>> containerOnDecom = getContainersOnDecomNodes();
@@ -237,7 +237,7 @@ public class TestEndpoints extends AbstractReconSqlDBTest {
         new ContainerWithPipeline(containerInfo, pipeline);
 
     mockScmClient = mock(
-        StorageContainerLocationProtocol.class, Mockito.RETURNS_DEEP_STUBS);
+        StorageContainerLocationInternalInterface.class, Mockito.RETURNS_DEEP_STUBS);
     StorageContainerServiceProvider mockScmServiceProvider = mock(
         StorageContainerServiceProviderImpl.class);
     when(mockScmServiceProvider.getPipeline(
@@ -284,7 +284,7 @@ public class TestEndpoints extends AbstractReconSqlDBTest {
             .addBinding(ContainerHealthSchemaManager.class)
             .addBinding(UtilizationEndpoint.class)
             .addBinding(ReconUtils.class, reconUtilsMock)
-            .addBinding(StorageContainerLocationProtocol.class, mockScmClient)
+            .addBinding(StorageContainerLocationInternalInterface.class, mockScmClient)
             .build();
 
     reconScm = (ReconStorageContainerManagerFacade)
