+++ hadoop-ozone/recon/src/test/java/org/apache/hadoop/ozone/recon/spi/impl/TestStorageContainerServiceProviderImpl.java
@@ -34,7 +34,7 @@ import org.apache.hadoop.hdds.conf.OzoneConfiguration;
 import org.apache.hadoop.hdds.protocol.proto.HddsProtos;
 import org.apache.hadoop.hdds.scm.pipeline.Pipeline;
 import org.apache.hadoop.hdds.scm.pipeline.PipelineID;
-import org.apache.hadoop.hdds.scm.protocol.StorageContainerLocationProtocol;
+import org.apache.hadoop.hdds.scm.protocol.StorageContainerLocationInternalInterface;
 import org.apache.hadoop.ozone.recon.ReconUtils;
 import org.apache.hadoop.ozone.recon.spi.StorageContainerServiceProvider;
 import org.junit.jupiter.api.BeforeEach;
@@ -55,15 +55,15 @@ public class TestStorageContainerServiceProviderImpl {
       @Override
       protected void configure() {
         try {
-          StorageContainerLocationProtocol mockScmClient = mock(
-              StorageContainerLocationProtocol.class);
+          StorageContainerLocationInternalInterface mockScmClient = mock(
+              StorageContainerLocationInternalInterface.class);
           ReconUtils reconUtils =  new ReconUtils();
           OzoneConfiguration conf = new OzoneConfiguration();
           conf.set(HddsConfigKeys.OZONE_METADATA_DIRS, testDir.getPath());
           pipelineID = PipelineID.randomId().getProtobuf();
           when(mockScmClient.getPipeline(pipelineID))
               .thenReturn(mock(Pipeline.class));
-          bind(StorageContainerLocationProtocol.class)
+          bind(StorageContainerLocationInternalInterface.class)
               .toInstance(mockScmClient);
           bind(StorageContainerServiceProvider.class)
               .to(StorageContainerServiceProviderImpl.class);
@@ -81,8 +81,8 @@ public class TestStorageContainerServiceProviderImpl {
   public void testGetPipelines() throws IOException {
     StorageContainerServiceProvider scmProvider =
         injector.getInstance(StorageContainerServiceProvider.class);
-    StorageContainerLocationProtocol scmClient =
-        injector.getInstance(StorageContainerLocationProtocol.class);
+    StorageContainerLocationInternalInterface scmClient =
+        injector.getInstance(StorageContainerLocationInternalInterface.class);
     scmProvider.getPipelines();
     verify(scmClient, times(1)).listPipelines();
   }
@@ -91,8 +91,8 @@ public class TestStorageContainerServiceProviderImpl {
   public void testGetPipeline() throws IOException {
     StorageContainerServiceProvider scmProvider =
         injector.getInstance(StorageContainerServiceProvider.class);
-    StorageContainerLocationProtocol scmClient =
-        injector.getInstance(StorageContainerLocationProtocol.class);
+    StorageContainerLocationInternalInterface scmClient =
+        injector.getInstance(StorageContainerLocationInternalInterface.class);
     Pipeline pipeline = scmProvider.getPipeline(pipelineID);
     assertNotNull(pipeline);
     verify(scmClient, times(1))
